Joined raw data
This rawSRDV1.csv consists of data that is in comma-separated Values (CSV). 
The File is a join of the two spreadsheets based upon the 'Sample ID'. 

The objective of unifying the two sheets is to perform Exploratory Data Analysis (EDA).

The CSV can be opened via MS Excel and any other spreadsheet software.